api/realtimequote/quotation/serializers.py
from rest_framework import serializers
from quotation.models import Quotation


class QuotationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Quotation
        fields = '__all__'