api/requirements/requirements.txt
celery[amqp]
django
djangorestframework
drf-spectacular
gunicorn
django-filter
django-cors-headers
django-celery-beat
django-celery-results
python-decouple
psycopg2-binary
requests
simple-settings
flower
pytest-django
pytest-mock
pytest-cov
